fix(backend): Refactored to more OO style
