+++ backend/src/index.ts
@@ -7,8 +7,6 @@
 import RESTServer from "./server";
 import mainRouter from "./routes/index";
 import {mongoose} from "./database/connection";
-import {GoogleProvider, NamedGoogleBackend} from "./types/terraform";
-import {rootBlock} from "./terraform/terraform";
 
 const server = new RESTServer();
 server.serve("/");
