+++ backend/src/terraform/googleBackend.ts
@@ -1,25 +1,45 @@
+import {model} from "mongoose";
 import CONFIG from "../config";
-import {GoogleBackend} from "../types/terraform";
+import {DatabaseModel, generateSchema} from "../types/database";
+import {GoogleBackend, named} from "../types/terraform";
 import {removeName} from "./util";
 
-export default (google: GoogleBackend) => {
-	return {
-		gcs: [removeName(google)].map(g => ({
-			bucket: g.bucket,
-			prefix: g.prefix
-		}))
-	};
-};
+export interface NamedGoogleBackend extends named<GoogleBackend, "gcs"> {}
+export class NamedGoogleBackend implements DatabaseModel<NamedGoogleBackend> {
+	name: "gcs";
+	constructor(bucket: string, prefix: string, location: string) {
+		this.name = "gcs";
+		this.bucket = bucket;
+		this.prefix = prefix;
+		this.location = location;
+	}
 
-export const toResource = (google: GoogleBackend) => {
-	const resource: any = {
-		google_storage_bucket: [{}]
-	};
-	resource.google_storage_bucket[0][CONFIG.TERRAFORM.BACKEND_BUCKET] = [
-		{
-			location: google.location,
-			name: google.bucket
-		}
-	];
-	return resource;
-};
+	toSchema() {
+		return generateSchema<NamedGoogleBackend>(this);
+	}
+	toModel() {
+		return model("GoogleBackend", this.toSchema());
+	}
+
+	toResource() {
+		const resource: any = {
+			google_storage_bucket: [{}]
+		};
+		resource.google_storage_bucket[0][CONFIG.TERRAFORM.BACKEND_BUCKET] = [
+			{
+				location: this.location,
+				name: this.bucket
+			}
+		];
+		return resource;
+	}
+
+	toJSON() {
+		return {
+			gcs: [removeName(this)].map(g => ({
+				bucket: g.bucket,
+				prefix: g.prefix
+			}))
+		};
+	}
+}
