+++ backend/src/types/terraform.ts
@@ -1,4 +1,8 @@
 import {Model, model, Schema} from "mongoose";
+import {NamedAwsBackend} from "../terraform/awsBackend";
+import {AwsProvider} from "../terraform/awsProvider";
+import {NamedGoogleBackend} from "../terraform/googleBackend";
+import {GoogleProvider} from "../terraform/googleProvider";
 import {
 	DatabaseModel,
 	generateSchema,
@@ -54,86 +58,13 @@ export class NamedRequiredProvider {
 	}
 }
 
-// -- -- -- //
-
-export interface GoogleProvider extends named<RequiredProvider, "google"> {
-	project?: string;
-	region?: string;
-	zone?: string;
-	credentials?: string;
-}
 export const isGoogleProvider = (
 	provider: NamedRequiredProvider
 ): provider is GoogleProvider => provider.name === "google";
 
-export class GoogleProvider
-	extends NamedRequiredProvider
-	implements GoogleProvider, DatabaseModel<GoogleProvider>
-{
-	name: "google";
-	constructor(
-		source: string,
-		version: string,
-		project: string | null = null,
-		region: string | null = null,
-		zone: string | null = null,
-		credentials: string | null = null
-	) {
-		super(source, version, "google");
-		this.name = "google";
-		if (project) this.project = project;
-		if (region) this.region = region;
-		if (zone) this.zone = zone;
-		if (credentials) this.credentials = credentials;
-	}
-
-	toSchema() {
-		return generateSchema<GoogleProvider>(this);
-	}
-
-	toModel() {
-		return model("GoogleProvider", this.toSchema());
-	}
-}
-
-// -- -- -- //
-
-export interface AwsProvider extends named<RequiredProvider, "aws"> {
-	region: string;
-	access_key: string;
-	secret_key: string;
-}
 export const isAwsProvider = (
 	provider: NamedRequiredProvider
 ): provider is AwsProvider => provider.name === "aws";
-
-export class AwsProvider
-	extends NamedRequiredProvider
-	implements DatabaseModel<AwsProvider>
-{
-	name: "aws";
-	constructor(
-		source: string,
-		version: string,
-		region: string,
-		access_key: string,
-		secret_key: string
-	) {
-		super(source, version, "aws");
-		this.name = "aws";
-		this.region = region;
-		this.access_key = access_key;
-		this.secret_key = secret_key;
-	}
-
-	toSchema() {
-		return generateSchema<AwsProvider>(this);
-	}
-
-	toModel() {
-		return model("AwsProvider", this.toSchema());
-	}
-}
 
 // --------------------------------Backend----------------------------------- //
 
@@ -144,27 +75,9 @@ export interface AwsBackend {
 	key: string;
 	region: string;
 }
-export interface NamedAwsBackend extends named<AwsBackend, "s3"> {}
 export const isAwsBackend = (
 	backend: namedTerraformBackend
-): backend is named<AwsBackend, "s3"> => backend.name === "s3";
-
-export class NamedAwsBackend implements DatabaseModel<NamedAwsBackend> {
-	name: "s3";
-	constructor(bucket: string, key: string, region: string) {
-		this.name = "s3";
-		this.bucket = bucket;
-		this.key = key;
-		this.region = region;
-	}
-
-	toSchema() {
-		return generateSchema<NamedAwsBackend>(this);
-	}
-	toModel() {
-		return model("AwsBackend", this.toSchema());
-	}
-}
+): backend is NamedAwsBackend => backend.name === "s3";
 
 // -- -- -- //
 
@@ -173,27 +86,9 @@ export interface GoogleBackend {
 	prefix: string;
 	location: string;
 }
-export interface NamedGoogleBackend extends named<GoogleBackend, "gcs"> {}
 export const isGoogleBackend = (
 	backend: namedTerraformBackend
-): backend is named<GoogleBackend, "gcs"> => backend.name === "gcs";
-
-export class NamedGoogleBackend implements DatabaseModel<NamedGoogleBackend> {
-	name: "gcs";
-	constructor(bucket: string, prefix: string, location: string) {
-		this.name = "gcs";
-		this.bucket = bucket;
-		this.prefix = prefix;
-		this.location = location;
-	}
-
-	toSchema() {
-		return generateSchema<NamedGoogleBackend>(this);
-	}
-	toModel() {
-		return model("GoogleBackend", this.toSchema());
-	}
-}
+): backend is NamedGoogleBackend => backend.name === "gcs";
 
 // -- -- -- //
